COMPLETED:
1. Removed all module dependencies;
2. Adjusted all modules to be optional (thereby removing
	the necessity of a "core");
3. Moved text traversal methods to `Utils.text`;
4. Updated documentation;
5. Added Bash command-line builder.

TODO:

1. Get more peer reviews;
2. Ponder ways to remove repetition;
3. Add Safari 1-2 to test browsers (Mac OS X);
4. Add Chrome 5+ to test browsers (???);
5. Ponder ways to support IE 4;
6. Ponder ways to truncate syntax.

TESTS:

(100% grade)

1. IE 5-9;
2. Firefox 1-12;
3. Opera 5-11;
4. Safari 3.1-5;
5. Chrome 1-4;
6. Netscape 6.10;

(incomplete)

7. IE 4 (0/64; SyntaxError);
8. Netscape 4.8 (0/64; SyntaxError).

NOTES:

* Sent an e-mail to the JSDaily team; a citation would be nice;
* Re-release may take place on HN tomorrow.

+++ compress.sh
@@ -7,9 +7,11 @@ compress ()
 {
 	INPUT="$1.js";
 	MIN_OUTPUT="${MIN}$1-min.js";
-	yui-compressor --type js --charset utf-8\
-		-o $MIN_OUTPUT ${FULL}$INPUT;
-	echo "$INPUT compressed";
+	if [ -f ${FULL}$INPUT ]; then
+		yui-compressor --type js --charset utf-8\
+			-o $MIN_OUTPUT ${FULL}$INPUT;
+		echo "$INPUT compressed";
+	fi;
 }
 
 compress_parts ()
@@ -25,10 +27,12 @@ compress_parts ()
 add_license ()
 {
 	NAME="$1-min.js";
-	INPUT="${MIN}$NAME";
 	LICENSE="${FULL}license.js";
-	cat $LICENSE $INPUT > "${MIN}temp";
-	mv "${MIN}temp" $INPUT;
+	INPUT="${MIN}$NAME";
+	if [ -f ${LICENSE} -a -f ${INPUT} ]; then
+		cat $LICENSE $INPUT > "${MIN}temp";
+		mv "${MIN}temp" $INPUT;
+	fi;
 }
 
 compress_builds ()
@@ -49,12 +53,12 @@ begin_parts ()
 		"metadata"
 		"raise"
 		"types"
-		"helpers"
 		"is"
 		"node"
 		"create"
 		"classes"
 		"traverse"
+		"text"
 		"select"
 	);
 	compress_parts "${PARTS[@]}";
@@ -63,11 +67,9 @@ begin_parts ()
 begin_builds ()
 {
 	BUILDS=(
-		"utils-core"
-		"utils-classes"
-		"utils-traverse"
-		"utils-select"
+		"utils-head"
 		"utils-complete"
+		"utils-build"
 	);
 	compress_builds "${BUILDS[@]}";
 }
